refactor: update import paths for domain configuration constants and types
