--- apps/web/src/store/config/constant.ts
@@ -1,12 +1,0 @@
-export const DOMAIN_CONFIG_NAMES = {
-    EXTRA_MENU: 'console:ext-menu',
-    SETTINGS: 'settings',
-    DORMANCY_WORKSPACE: 'identity:dormancy:workspace',
-    ANOMALY_DETECTION_CONFIGURATION: 'anomaly_detection_configuration',
-} as const;
-
-export const SHARED_CONFIG_NAMES = {
-    TASK_LANDING: 'console:task-management:landing',
-    TASK_TEMPLATE: 'console:task-management:template',
-} as const;
-
